+++ zh/guide_cloud/general/changelog.md
@@ -45,7 +45,7 @@ Github: https://github.com/SwanHubX/SwanLab
 
 ## v0.7.4 - 2025.12.14
 
-发布 SwanLab Kubernetes版本，部署方式见[此文档](/guide_cloud/self_host/kubernetes-deploy.md)
+发布 SwanLab Kubernetes版本，部署方式见[此文档](/guide_cloud/self_host/kubernetes/deploy.md)
 
 ![alt](https://swanlab-docs-1301372061.cos.ap-beijing.myqcloud.com/assets/zh/guide_cloud/self_host/kubernetes/logo.png)
 
@@ -315,9 +315,9 @@ Github: https://github.com/SwanHubX/SwanLab
 
 ## v0.5.0 - 2025.3.12
 
-![logo](https://swanlab-docs-1301372061.cos.ap-beijing.myqcloud.com/assets/zh/guide_cloud/self_host/docker-deploy/swanlab-docker.jpg)
+![logo](https://swanlab-docs-1301372061.cos.ap-beijing.myqcloud.com/assets/zh/guide_cloud/self_host/docker/docker-deploy/swanlab-docker.jpg)
 
-**🎉🎉SwanLab私有化部署（社区版）现已重磅发布！！**[部署文档](/guide_cloud/self_host/docker-deploy.md)
+**🎉🎉SwanLab私有化部署（社区版）现已重磅发布！！**[部署文档](/guide_cloud/self_host/docker/deploy.md)
 
 **🚀新增功能**
 - `swanlab.init`新增参数`callbacks`，支持在初始化时注册回调函数，以支持各式各样的自定义插件类